indexing refactor and passing in examples
