+++ docugami_kg_rag/config/__init__.py
@@ -27,5 +27,7 @@ LOCAL_LLM_CACHE_DB_FILE = os.environ.get("LOCAL_LLM_CACHE", "/tmp/docugami/.lang
 os.makedirs(Path(LOCAL_LLM_CACHE_DB_FILE).parent, exist_ok=True)
 set_llm_cache(SQLiteCache(database_path=LOCAL_LLM_CACHE_DB_FILE))
 
+EXAMPLES_PATH = Path(__file__).parent.parent / "green_examples"
+
 DEFAULT_USE_REPORTS = False
 AGENT_MAX_ITERATIONS = 5
