+++ docugami_kg_rag/green_examples/describe_document_set_examples.yaml
@@ -1,0 +1,45 @@
+- summaries: |
+    ****************
+    DOCUMENT SUMMARY:
+
+    On November 21, 2007, at 0001 UTC, a Boeing B737-400, registered under DCA08WA016, experienced an incident near Jakarta, Indonesia. The aircraft, operated by Batavia Airlines, was forced to return to Jakarta-Sukarno-Hatta airport
+    following a loss of hydraulic pressure. Minor damage was reported, including a missing upper wing panel and damaged adjacent hydraulic tubing. The incident occurred shortly after routine maintenance, where some fasteners were not
+    reinstalled correctly. There were no injuries to the crew or passengers, but the aircraft sustained minor damage. The flight was conducted under non-U.S., commercial regulations.
+
+    The National Transportation Safety Committee of Indonesia is leading the investigation, with the Ministry of Communications of the Republic of Indonesia providing oversight. The aircraft was not amateur-built, had its last
+    inspection noted without specifying the date, and was certified for commercial operations without any operating certificates held. Meteorological conditions at the time of the incident were visual, with the flight plan filed
+    under IFR conditions. The departure point was Jakarta, with the destination unknown.
+
+    Airport information indicates the use of a runway compatible with IFR approaches, and wreckage examination confirmed the minor damage to the aircraft without any fire or explosion. There were no injuries on the ground. The
+    investigation docket is managed by the NTSB, with William English serving as the Investigator in Charge (IIC). Public access to investigation dockets prior to June 1, 2009, is available through NTSB's Record Management Division,
+    with subsequent dockets accessible online at the NTSB's official docket management system.
+
+    ****************
+    DOCUMENT SUMMARY:
+
+    National Transportation Safety Board Aviation Accident Factual Report details an accident involving a Cirrus Design Corp. SR-20, registration N141CD, which occurred on November 28, 2007, at 1800 UTC near Jagel, Germany. The aircraft
+    was destroyed after striking power lines and impacting terrain during an approach for landing, resulting in the fatal injury of the private pilot, the sole occupant. The flight, originating in the Netherlands and destined for Sylt,
+    Germany, was conducted under Part 91: General Aviation as a ferry flight. Instrument meteorological conditions prevailed at the time of the accident.
+
+    The aircraft, manufactured by Cirrus Design Corporation with serial number 1032, sustained substantial damage. The investigation is under the jurisdiction of the German Federal Bureau of Aircraft Accidents Investigation (BFU),
+    located in Braunschweig, Germany. Information regarding the pilot's certification, ratings, and medical certification is included but specifics are not detailed in the summary. The aircraft and owner/operator information,
+    along with meteorological conditions, flight plan, airport information, and wreckage and impact information, are documented, highlighting the substantial aircraft damage and the fatal injury to the pilot.
+
+    The report mentions that the investigation docket is maintained by the NTSB, with Investigator in Charge (IIC) David C. Bowling and additional participation from Karsten Severin of the German BFU. The accident and incident
+    dockets are available for public access, providing archival information for NTSB investigations.
+
+    ****************
+    DOCUMENT SUMMARY:
+
+    The National Transportation Safety Board (NTSB) issued a final report for aviation accident number SEA08CA039, which occurred in Yelm, WA, on November 24, 2007, at 1030 PST. The accident involved a Piper PA-18A-135, registration
+    N3250B, which sustained substantial damage during landing on a gravel river bar. The pilot, a 37-year-old male with a private certificate and 228 hours of total flight time, failed to maintain directional control during the
+    landing roll, resulting in the aircraft impacting a log adjacent to the landing area. There were no mechanical anomalies found during the post-accident inspection. The pilot was the only person on board and sustained no injuries.
+    The flight was conducted under Part 91: General Aviation for personal reasons, with no flight plan filed.
+
+    The probable cause of the accident was identified as the pilot's failure to maintain directional control of the tailwheel-equipped airplane during the landing roll. Contributing factors included the rough/uneven terrain and
+    the presence of a log. The report also provided detailed information on the pilot's qualifications, aircraft data, meteorological conditions, and administrative details, including the investigator in charge, Joshua D Cawthra.
+    The NTSB's investigation aims to improve transportation safety by determining the causes of accidents and issuing recommendations to prevent future incidents.
+  docset_name: NTSB Accident Report
+  description: |
+    This document type is a detailed report issued by the National Transportation Safety Board (NTSB) regarding aviation accidents or incidents. It includes information about the circumstances, causes, and findings of the accident
+    or incident, as well as details about the aircraft, individuals involved, and the investigation process.
