+++ docugami_kg_rag/green_examples/summarize_chunk_examples.yaml
@@ -1,0 +1,45 @@
+- format: text
+  contents: |
+    This Agreement shall be governed by and construed in accordance with the laws of the State of [State],
+    without regard to its conflict of laws principles.
+  summary: Governing law and jurisdiction determined by the laws of [State].
+
+- format: text
+  contents: |
+    Each party shall indemnify, defend and hold harmless the other party and its officers, directors, employees,
+    agents, affiliates, successors and assigns from and against any and all claims, losses, damages, liabilities,
+    penalties, judgments, costs and expenses (including reasonable attorneys' fees) arising out of or resulting
+    from the indemnifying party's gross negligence or willful misconduct in connection with this Agreement.
+  summary: Mutual indemnification for negligence or misconduct.
+
+- format: text
+  contents: |
+    Either party may terminate this Agreement upon thirty (30) days' written notice to the other party if the
+    other party breaches any material term or condition of this Agreement and fails to cure such breach within
+    such thirty (30) day period.
+  summary: Termination for material breach with 30-day cure period.
+
+- format: text
+  contents: |
+    The term of this Agreement shall commence on the Effective Date and continue in full force and effect for
+    an initial term of one (1) year, and shall automatically renew for successive one (1) year terms, unless
+    either party provides written notice of its intent not to renew at least thirty (30) days prior to the
+    expiration of the then-current term.
+  summary: Initial term of one year with automatic renewal, unless 30-day notice given.
+
+- format: text
+  contents: |
+    All notices, requests, demands, and other communications under this Agreement shall be in writing and
+    shall be deemed to have been given on the date of service if served personally, or three (3) days after
+    mailing if mailed, by first class mail, registered or certified, postage prepaid, and addressed to the
+    parties at the addresses specified in this Agreement or at such other address as a party may designate
+    by notice given as herein provided.
+  summary: Written notices deemed given on date of service or three days after mailing.
+
+- format: text
+  contents: |
+    The parties acknowledge that the relationship between them is that of independent contractors. Nothing
+    contained in this Agreement shall be construed as creating any agency, partnership, joint venture or
+    other form of joint enterprise, employment or fiduciary relationship between the parties, and neither
+    party shall have authority to contract for or bind the other party in any manner whatsoever.
+  summary: Parties as independent contractors, not agents, employees, or partners.
