+++ docugami_kg_rag/green_examples/summarize_document_examples.yaml
@@ -1,0 +1,65 @@
+- format: text
+  contents: |
+    INITIAL STATEMENT OF WORK (SOW)
+    The purpose of this SOW is to describe the Services that Company will initially provide to Zinga, Inc. the “Client”) under the terms and conditions of the Services Agreement entered into between the parties on March 15, 2021 (the “Agreement”). Additional terms and conditions may be set forth in this SOW, and additional SOWs that may be subsequently signed.
+    To the extent the terms and conditions of this SOW are inconsistent with those of the Agreement, the terms of this SOW will control with respect to the Services described herein. Capitalized terms used herein shall have the same meaning as those used in the Agreement.
+    This SOW is an attachment to and is incorporated by this reference into the Agreement as if fully set forth therein and made a part thereof. This SOW, together with the Agreement, represents the complete and total understanding of the parties regarding the Services to be provided by Company hereunder.
+    1. Services Period .
+    Unless otherwise terminated earlier in accordance with the terms of the Agreement, the Services will end upon the completion of the Services by Company, which in no event shall be later than March 15, 2022 , and payment by client.
+    2. Services.
+    The Services outlined in this SOW reflect the initial understanding of Client’s desired Deliverable and Company reserves the right to alter the manner in which Services are provided if Company determines different services are necessary to complete the Deliverable. Company shall provide the following Services to the Client:
+    +----------------------+----------+-------------+---------------+
+    | Description          | Duration | Hourly Rate | Subtotal      |
+    +----------------------+----------+-------------+---------------+
+    | SOC Analyst          | 20 weeks | $62.60      | $520,832.00   |
+    +----------------------+----------+-------------+---------------+
+    | Project Engineer III | 20 weeks | $88.92      | $400,384.00   |
+    +----------------------+----------+-------------+---------------+
+    | Software Engineer IV | 20 weeks | $105.00     | $100,200.00   |
+    +----------------------+----------+-------------+---------------+
+    | Network Engineer III | 20 weeks | $88.92      | $105,953.60   |
+    +----------------------+----------+-------------+---------------+
+    | ESTIMATED TOTAL      |          |             | $1,126,369.60 |
+    +----------------------+----------+-------------+---------------+
+    3. Deliverables.
+    Company shall deliver the following Deliverable:
+    A. Conduct a one-week Discovery session with Zinga to understand needs, identify stakeholders and research areas.
+    B. Conduct ten 45-minute User Research sessions with Zinga Customers
+    C. Conduct three one-week Design Sprints to ideate possible product solutions
+    D. Document Identified Customer Personas , Journey Maps & Design Sprint Ideas for Zinga
+    4. Fees and Terms.
+    Total Amount : $ 1,126,369.60 Start Date : 03/03/2021 End Date: 06/01/2021 If the Client desires to initiate changes to the SOW, it shall submit to Company a written request to do so. The request will set forth the nature of the Company’s proposed changes to the SOW. Contractor shall complete and return to the Company a written document (“ Change Order ”) setting forth (i) a written description of the changes to the SOW, (ii) any changes to the schedule, (iii) any changes or additions to the Deliverables and (iv) any changes or additions to the fees. A Change Order will be binding only if signed by both parties. Any and all Change Orders will be governed by the terms and conditions set forth in this Agreement and are hereby incorporated by this reference. Any additional Deliverables described in the Change Order will be subject to the Payment provisions as described in Section 7 of the Agreement.
+    5. Hours of Operation
+    • Engineering:
+    • Monday - Friday 6AM-6PM PST
+    • engineering@digitalmemex.com
+    • ( 206 ) 684 - 8889
+    • Support:
+    • 24 x 7 x 365
+    • support@digitalmemex.com
+    • ( 206 ) 684 - 8888
+    6 . Travel.
+    If the Deliverable will be completed according to a fixed price SOW and travel is required, then Client must sign the Agreement and SOW and pay required Deposit at least 30 days prior to proposed travel. If the Agreement and SOW are not signed and Deposit is not received at least 30 days prior to proposed travel, Company reserves the right to amend the SOW estimate solely with regard to travel and lodging expenses so that the SOW estimate reflects actual costs incurred by Company.
+    [ Signature Page Follows]
+    Accepted and agreed to as of the SOW Effective Date by the authorized representative of each party.
+    +----------------------------+--------+-------------------------------------+
+    | Zinga, Inc.                |        | Digital Memories Corporation , Inc. |
+    +----------------------------+--------+-------------------------------------+
+    | Signature:                 | acelin | Signature: Ye                       |
+    +----------------------------+--------+-------------------------------------+
+    | Print Name : Aziz Khan     |        | Print Name : Morris Minor           |
+    +----------------------------+--------+-------------------------------------+
+    | Print Title : COO          |        | Print Title : CEO                   |
+    +----------------------------+--------+-------------------------------------+
+    | Date: 06/01/2021           |        | Date: 06/01/2021                    |
+    +----------------------------+--------+-------------------------------------+
+    | Email: aziz.khan@zinga.com |        | Email: morris@digitalmemex.com      |
+    +----------------------------+--------+-------------------------------------+
+  summary: |
+    This document is an initial Statement of Work (SOW) between Company and Zinga, Inc. It outlines the services that Company will provide to Zinga under the Services Agreement.
+    The SOW is an attachment to the Agreement and represents the complete understanding of the parties regarding the services to be provided. The Services Period will end no later
+    than March 15, 2022. The Services include SOC Analyst, Project Engineer III, Software Engineer IV, and Network Engineer III, with respective durations and hourly rates.
+    The estimated total cost of the Services is $1,126,369.60. The Deliverables include a Discovery session, User Research sessions, Design Sprints, and documentation of 
+    Customer Personas, Journey Maps, and Design Sprint Ideas. Changes to the SOW can be initiated through a written request and a Change Order must be signed by both parties to be binding.
+    The hours of operation for Engineering and Support are specified. Travel expenses will be incurred if required and must be accounted for in the SOW. 
+    The document includes signature and contact information for both parties.
