+++ docugami_kg_rag/state_models.py
@@ -1,9 +1,30 @@
-from dataclasses import dataclass, field
+from dataclasses import dataclass
+from pathlib import Path
+
+from dataclasses import field
 from typing import List
 
-
 from langchain.storage.in_memory import InMemoryStore
-from docugami_kg_rag.helpers.reports import ReportDetails
+
+
+@dataclass
+class ReportDetails:
+    id: str
+    """ID of report."""
+
+    name: str
+    """Name of report."""
+
+    local_xlsx_path: Path
+    """Local path to XLSX of the report."""
+
+    retrieval_tool_function_name: str
+    """Function name for retrieval tool e.g. sql_query_earnings_calls."""
+
+    retrieval_tool_description: str
+    """
+    Description of retrieval tool e.g. Runs a SQL query over the REPORT_NAME report, 
+    represented as the following SQL Table... etc."""
 
 
 @dataclass
