+++ index.py
@@ -2,7 +2,7 @@ import sys
 import typer
 from docugami import Docugami
 
-from docugami_kg_rag.helpers.indexing import index_docset
+from docugami_kg_rag.indexing import index_docset
 
 docugami_client = Docugami()
 
@@ -42,6 +42,7 @@ if __name__ == "__main__":
     if sys.gettrace():
         # This code will only run if a debugger is attached
         # index_docset(docset_id="tjwrr2ekqkc3", name="SEC 10Q Reports", overwrite=True)
-        index_docset(docset_id="uh31voknl1p6", name="NTSB Accident Report", overwrite=True)
+        # index_docset(docset_id="uh31voknl1p6", name="NTSB Accident Report", overwrite=True)
+        index_docset(docset_id="ue548ls3yjqu", name="City of Anacortes Contractor Agreement")
     else:
         app()
